+++ src/constants/index.ts
@@ -1,0 +1,14 @@
+export const CATEGORIES = {
+  ALL: '전체',
+  KOREAN: '한식',
+  CHINESE: '중식',
+  JAPANESE: '일식',
+  WESTERN: '양식',
+  ASIAN: '아시안',
+  ETC: '기타',
+} as const;
+
+export const SORT_OPTIONS = {
+  NAME: '이름순',
+  DISTANCE: '거리순',
+};
